test(pipeline/gate): remove tautological/subsumed tests, relocate misplaced ones
- Delete test_ungated_tools_equals_free_range_tools (tautological:
  UNGATED_TOOLS is literally defined as FREE_RANGE_TOOLS alias)
- Delete test_gated_tools_does_not_contain_run_recipe (subsumed by
  test_gated_tools_contains_expected_names which uses full set equality)
- Move test_headless_tools_contains_expected_names and
  test_free_range_tools_contains_expected_names to
  tests/core/test_type_constants.py — they test core-layer constants
  and belong in the core layer-mirroring test directory
