+++ tests/pipeline/test_gate.py
@@ -165,35 +165,6 @@ def test_gate_imports_only_from_core():
                     )
 
 
-def test_gated_tools_does_not_contain_run_recipe():
-    from autoskillit.pipeline.gate import GATED_TOOLS
-
-    assert "run_recipe" not in GATED_TOOLS
-
-
-def test_headless_tools_contains_expected_names():
-    from autoskillit.core.types import HEADLESS_TOOLS
-
-    assert HEADLESS_TOOLS == {"test_check"}
-
-
-def test_free_range_tools_contains_expected_names():
-    from autoskillit.core.types import FREE_RANGE_TOOLS
-
-    assert FREE_RANGE_TOOLS == {
-        "open_kitchen",
-        "close_kitchen",
-        "disable_quota_guard",
-        "reload_session",
-    }
-
-
-def test_ungated_tools_equals_free_range_tools():
-    from autoskillit.core.types import FREE_RANGE_TOOLS, UNGATED_TOOLS
-
-    assert UNGATED_TOOLS == FREE_RANGE_TOOLS
-
-
 def test_all_tool_sets_disjoint_and_complete():
     from autoskillit.core.types import HEADLESS_TOOLS
     from autoskillit.pipeline.gate import GATED_TOOLS, UNGATED_TOOLS
